+++ CHANGES.rst
@@ -10,6 +10,8 @@ https://zope.readthedocs.io/en/2.13/CHANGES.html
 4.8.7 (unreleased)
 ------------------
 
+- Only set response header Content-Type as text/html on exception views when the response has content.
+
 - Update dependencies to the latest releases for each supported Python version.